Create mydns.c
This is the final mydns program in C. This can be run in a Linux environment. In this case, we run it on our school server ocelot.aul.fiu.edu by using Putty and emacs. Also, we used WinScp to download the file and upload it to the FIU server.

Co-Authored-By: ipromiseshes18 <67117981+ipromiseshes18@users.noreply.github.com>
Co-Authored-By: yhan00 <59639002+yhan00@users.noreply.github.com>